Add rspec and chefspec tests
This does change a few things internally.

Validator now does not attempt to access Chef objects directly and instead expects hashes to be sent to it. One of the annoying pieces here is around .
This was done both to make testing simpler and to make it possible to use the gem *outside* the Chef context, such as I have the data structure already from Chef and just want to validate it in Ruby